Added an option to skip apt-get update & upgrade, removed some osx code

+++ init/10_ubuntu.sh
@@ -31,31 +31,34 @@ EOF
 fi
 
 # Update APT.
-e_header "Updating APT"
-sudo apt-get -qq update
-sudo apt-get -qq upgrade
+read -N 1 -t 15 -p "Do apt-get update & upgrade? [y/N] " apt_get_update; echo
+if [[ "$apt_get_update" =~ [Yy] ]]; then
+    e_header "Updating APT"
+    sudo apt-get -qq update
+    sudo apt-get -qq upgrade
 
-# Install APT packages.
-packages=(
-  build-essential libssl-dev
-  git-core
-  tree sl id3tool
-  nmap telnet
-  htop
-)
+    # Install APT packages.
+    packages=(
+      build-essential libssl-dev
+      git-core
+      tree sl id3tool
+      nmap telnet
+      htop
+    )
 
-list=()
-for package in "${packages[@]}"; do
-  if [[ ! "$(dpkg -l "$package" 2>/dev/null | grep "^ii  $package")" ]]; then
-    list=("${list[@]}" "$package")
-  fi
-done
+    list=()
+    for package in "${packages[@]}"; do
+      if [[ ! "$(dpkg -l "$package" 2>/dev/null | grep "^ii  $package")" ]]; then
+        list=("${list[@]}" "$package")
+      fi
+    done
 
-if (( ${#list[@]} > 0 )); then
-  e_header "Installing APT packages: ${list[*]}"
-  for package in "${list[@]}"; do
-    sudo apt-get -qq install "$package"
-  done
+    if (( ${#list[@]} > 0 )); then
+      e_header "Installing APT packages: ${list[*]}"
+      for package in "${list[@]}"; do
+        sudo apt-get -qq install "$package"
+      done
+    fi
 fi
 
 # Install Git Extras
